arch/arm/stm32u5: Move reusable core support to common
Move the STM32U5 startup, interrupt, GPIO, EXTI, serial, heap, idle, and
timer interrupt implementations into arch/arm/src/common/stm32.

This prepares the Cortex-M33 support for reuse by compatible STM32 families.
