+++ arch/arm/src/common/stm32/CMakeLists.txt
@@ -34,6 +34,24 @@ set(SRCS)
 list(APPEND SRCS stm32_waste.c)
 list(APPEND SRCS stm32_uid.c)
 
+if(CONFIG_ARCH_CHIP_STM32U5)
+  list(
+    APPEND
+    SRCS
+    stm32_allocateheap_m33_u5.c
+    stm32_exti_gpio_m33_u5.c
+    stm32_gpio_m33_u5.c
+    stm32_irq_m33_u5.c
+    stm32_lowputc_m33_u5.c
+    stm32_serial_m33_u5.c
+    stm32_start_m33_u5.c
+    stm32_timerisr_m33_u5.c)
+
+  if(NOT CONFIG_ARCH_IDLE_CUSTOM)
+    list(APPEND SRCS stm32_idle_m33_u5.c)
+  endif()
+endif()
+
 if(CONFIG_STM32_PULSECOUNT)
   list(APPEND SRCS stm32_pulsecount.c)
 endif()
