+++ arch/arm/src/common/stm32/hardware/stm32_gpio.h
@@ -27,7 +27,9 @@
  * Included Files
  ****************************************************************************/
 
-#if defined(CONFIG_STM32_HAVE_IP_GPIO_M0_V1)
+#if defined(CONFIG_ARCH_CHIP_STM32U5)
+#  include "hardware/stm32_gpio_m33_u5.h"
+#elif defined(CONFIG_STM32_HAVE_IP_GPIO_M0_V1)
 #  include "hardware/stm32_gpio_v2_m0.h"
 #elif defined(CONFIG_STM32_STM32L15XX)
 #  include "hardware/stm32l15xxx_gpio.h"
