+++ arch/arm/src/common/stm32/hardware/stm32_syscfg.h
@@ -30,7 +30,9 @@
 #include <nuttx/config.h>
 #include "chip.h"
 
-#if defined(CONFIG_ARCH_CHIP_STM32F0)
+#if defined(CONFIG_ARCH_CHIP_STM32U5)
+#  include "hardware/stm32_syscfg_m33_u5.h"
+#elif defined(CONFIG_ARCH_CHIP_STM32F0)
 #  include "hardware/stm32f0_syscfg.h"
 #elif defined(CONFIG_ARCH_CHIP_STM32L0)
 #  include "hardware/stm32l0_syscfg.h"
