+++ arch/arm/src/common/stm32/stm32_rcc.h
@@ -27,6 +27,12 @@
  * Included Files
  ****************************************************************************/
 
+#if defined(CONFIG_ARCH_CHIP_STM32U5)
+
+#  include "stm32_rcc_m33_u5.h"
+
+#else
+
 #include <nuttx/config.h>
 
 #include "arm_internal.h"
@@ -224,4 +230,5 @@ void stm32_rcc_disablelsi(void);
 #endif
 #endif /* __ASSEMBLY__ */
 
+#endif /* CONFIG_ARCH_CHIP_STM32U5 */
 #endif /* __ARCH_ARM_SRC_COMMON_STM32_STM32_RCC_H */
