+++ arch/arm/src/stm32u5/Make.defs
@@ -33,15 +33,9 @@ include common/stm32/Make.defs
 # Required STM32U5 files
 
 CHIP_ASRCS  =
-CHIP_CSRCS += stm32_allocateheap.c stm32_exti_gpio.c stm32_gpio.c
-CHIP_CSRCS += stm32_irq.c stm32_lowputc.c stm32_rcc.c stm32_i2c.c
-CHIP_CSRCS += stm32_serial.c stm32_start.c
+CHIP_CSRCS += stm32_rcc.c stm32_i2c.c
 CHIP_CSRCS += stm32_spi.c stm32_lse.c stm32_lsi.c stm32u5xx_rcc.c
-CHIP_CSRCS += stm32_pwr.c stm32_tim.c stm32_flash.c stm32_timerisr.c
-
-ifneq ($(CONFIG_ARCH_IDLE_CUSTOM),y)
-CHIP_CSRCS += stm32_idle.c
-endif
+CHIP_CSRCS += stm32_pwr.c stm32_tim.c stm32_flash.c
 
 ifeq ($(CONFIG_TIMER),y)
 CHIP_CSRCS += stm32_tim_lowerhalf.c
